% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{USD amount in EUR}
\usage{
convert_usd_to_eur(usd_amount)
}
\arguments{
\item{usd_amount}{The amount in USD to be converted to EUR.}
}
\value{
Returns the USD amount in EUR as a number.
}
\description{
Looks up the most recent USD/EUR exchange rate and computes the provided USD amount in EUR.
}
\examples{
convert_usd_to_eur(100)
}
