% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Create Euro format}
\usage{
eur(x)
}
\arguments{
\item{x}{The number to be formatted as Euro.}
}
\value{
Returns a string in Euro format.
}
\description{
Converts a number to a string using the Euro sign, rounding up to 2 digits and using the comma as the separator.
}
\examples{
eur(100)
eur(-100000.34561)

}
\seealso{
\code{\link{uneur}}
}
