% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{USD/EUR exchange rate}
\usage{
eurusd()
}
\value{
Returns the exchange rate as a number.
}
\description{
Looks up the most recent USD/EUR exchange rate via an API call and returns the rate as a number.
}
\examples{
eurusd()

}
\seealso{
\code{\link{eurusds}}
}
