% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{USD/EUR exchange rates}
\format{Returns a \code{data.table}.
\itemize{
    \item \code{date}: The day for which the exchange rate was valid.
    \item \code{exchange_rate}: Daily exchange rate.
}}
\usage{
eurusds(date_from, date_to)
}
\arguments{
\item{date_from}{The start of the range.}

\item{date_to}{The end of the range.}
}
\value{
Returns the daily exchange rates.
}
\description{
Looks up the daily USD/EUR exchange rate via an API call for the specified date range.
}
\examples{
eurusds("2019-05-01", "2019-05-24")

}
\seealso{
\code{\link{eurusd}}
}
