% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Reverse Euro format}
\usage{
uneur(x)
}
\arguments{
\item{x}{The string in Euro format to be transformed to number.}
}
\value{
Returns a number.
}
\description{
Converts a string in Euro format to number.
}
\examples{
uneur("€100,000.34")

}
\seealso{
\code{\link{eur}}
}
