% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{USD/EUR exchange rates}
\format{Returns a \code{data.table}.
\itemize{
    \item \code{date}: The day for which the exchange rate was valid.
    \item \code{exchange_rate}: Daily exchange rate.
}}
\usage{
eurusds(date_from, date_to, last_x_days = NULL)
}
\arguments{
\item{date_from}{The start of the range.}

\item{date_to}{The end of the range.}

\item{last_x_days}{Return the exchange rates for the last X days compared to \code{date_to}.}
}
\value{
Returns the daily exchange rates.
}
\description{
Looks up the daily USD/EUR exchange rate via an API call for the specified date range.
}
\examples{
# Specific date range
eurusds("2019-05-01", "2019-05-24")

# Last 45 days from a specific date
eurusds(date_to = "2019-05-01", last_x_days = 45)

# Last 45 days from today
eurusds(last_x_days = 45)

}
\seealso{
\code{\link{eurusd}}
}
